bug fixes & added 32bit counter support
git-svn-id: svn+ssh://repo.cacti.net/var/data/svnroot/cacti/trunk/cacti@257 860744bd-22fc-0310-8c96-e9fe5004b5ca

+++ dev/poller/global.h
@@ -26,7 +26,7 @@ typedef struct threads_struct{
 
 void *poller(void *thread_args);
 void *sig_handler();
-int snmp_get(char *snmp_host, char *snmp_comm, char *snmp_oid, int who);
+unsigned long long int snmp_get(char *snmp_host, char *snmp_comm, char *snmp_oid, int who);
 int lock;
 int sighup;
